+++ invenio_rdm_records/oai.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) 2021 Graz University of Technology.
-# Copyright (C) 2021-2023 CERN.
+# Copyright (C) 2021-2024 CERN.
 #
 # Invenio-RDM-Records is free software; you can redistribute it and/or modify
 # it under the terms of the MIT License; see LICENSE file for more details.
@@ -148,6 +148,7 @@ class OAIRecordSearch(RecordsSearch):
         default_filter = [
             dsl.Q("exists", field="pids.oai.identifier"),
             dsl.Q("term", **{"access.record": "public"}),
+            dsl.Q("term", **{"is_deleted": "false"}),
         ]