+++ tests/resources/conftest.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2020 CERN.
+# Copyright (C) 2020-2024 CERN.
 # Copyright (C) 2020 Northwestern University.
 # Copyright (C) 2021 TU Wien.
 #
@@ -35,18 +35,18 @@ def link(url):
 @pytest.fixture(scope="function")
 def app_with_allowed_edits(running_app):
     """Allow editing of published records."""
-    running_app.app.config[
-        "RDM_LOCK_EDIT_PUBLISHED_FILES"
-    ] = lambda service, identity, record=None: False
+    running_app.app.config["RDM_LOCK_EDIT_PUBLISHED_FILES"] = (
+        lambda service, identity, record=None: False
+    )
     return running_app
 
 
 @pytest.fixture(scope="function")
 def app_with_deny_edits(running_app):
     """Deny editing of published records."""
-    running_app.app.config[
-        "RDM_LOCK_EDIT_PUBLISHED_FILES"
-    ] = lambda service, identity, record=None: True
+    running_app.app.config["RDM_LOCK_EDIT_PUBLISHED_FILES"] = (
+        lambda service, identity, record=None: True
+    )
 
 
 def init_file(client, recid, key, headers):