+++ CHANGES.rst
@@ -1,6 +1,6 @@
 
 ..
-    Copyright (C) 2019-2023 CERN.
+    Copyright (C) 2019-2024 CERN.
     Copyright (C) 2019 Northwestern University.
 
 
@@ -10,6 +10,73 @@
 
 Changes
 =======
+
+Version 7.1.1 (released 2024-02-19)
+
+- communities: add CommunityParentComponent
+
+Version 7.1.0 (released 2024-02-19)
+
+- mappings: change "dynamic" values to string
+- requests: change default removal reason to spam
+- mappings: add keyword field to ``funding.award.number``
+- files: fixed infinite spinning wheel on error
+- datacite: added config for funders id priority
+- datacite: updated schema
+- mapping: add community children
+
+Version 7.0.0 (released 2024-02-16)
+
+- services: update community components
+- installation: bump invenio-communities
+- mappings: denormalize communities in records
+- systemfields: fix docstrings
+- requests: add check on parent community on accept
+- community selection: small ui fixes
+
+Version 6.2.1 (released 2024-02-11)
+
+- requests: add record to parent community
+
+Version 6.2.0 (released 2024-02-09)
+
+- tests: make deleted file fetching deterministic
+- deposit: change upload workflow for styled communities
+- deposit: indicate if community selection modal is used for initial submission
+- deposit: add community.theme.enabled to selection modal
+- installation: bump invenio-communities version
+
+Version 6.1.1 (released 2024-02-05)
+
+- oai: exclude deleted records from search
+- models: add bucket_id index
+- serializers: fix DataDownload missing mimetype
+
+Version 6.1.0 (released 2024-02-01)
+
+- Add CSV records serializer
+
+Version 6.0.0 (released 2024-01-31)
+
+- installation: bump dependencies
+- installation: pin commonmeta-py
+
+Version 5.1.1 (released 2024-01-30)
+
+- Custom field ui: fix deserializing for primitive types
+
+Version 5.1.0 (released 2024-01-29)
+
+- pids: restore required PIDs on publish
+- schema: add dataset specific fields to jsonld
+
+Version 5.0.0 (2024-01-16)
+
+- communities: utilize community theming mechanism
+
+Version 4.43.2 (2023-01-16)
+
+- dependencies: pin commonmeta-py
 
 Version 4.43.1 (2023-12-12)